requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
Flask==2.2.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.14
Werkzeug==2.2.3
youtube-transcript-api==0.5.0
zipp==3.15.0
